Update 02-art.tex
Added comments, corrected mispelling, and revised phrasing. I am not convinced of the overall structure, since the organization of the chapter is not clear and the relationship of the literature with what you have done is only in few points pointed out. This can be improved ad possibly a structure and rationale for it can be given at the beginning of the chapter.