Update 05-modeling.tex
EDITED and COMMENTED .This chapter contains parts that are general enough to become part of the introduction, and frame the whole thesis, other parts that are quite specific to a paper (and it is worth to state explicitly that they have been published), other parts that could be related to other parts of the thesis. I have put comments. Let's try to have something more similar to a chapter of a thesis, integrated in the thesis flow, than just a paper put here out of the blue.